+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/internal/StyleSlot.kt
@@ -53,7 +53,7 @@ internal fun StyleSlot(
       factory = {
         StyleSlotNode(name)
       },
-      update = { }
+      update = {}
     ) {
       content?.invoke()
     }